Add commands to backup and restore the database
We will create a cronjob that will be responsible for running the
backup_database task for us (probably once a day).

Also, add a command handler similar to dbadmin's one that will be
responsible for handling commands on the server (e.g. running the
server itself, running the backup/restore task, etc)

Change-Id: I85a871db10598253940cd37914917c9aee6b5df1