Fix backups failing due to an unremoved lockfile
When duplicity fails in unpredicted situations (e.g. the power is shut down
suddenly) it can leave a lockfile behind, and that can make any future backup
attempts fail to.  Check if that was the reason of the failure and, if the
lockfile is older than 3h remove it and try again.

Change-Id: I8ce73fa082c0c697a3bf3cc10089e752582a4f88