Standardized the client resource output and allowed changing document
The client resource was providing different outputs according to the
method that was used. Now, all methods will return the same output
with help of the method _dump_client.

A change was made also in the sale resource to allow modifying of the
client document without changing the whole client.

The method get_data now returns staffId of ease of use.

Fixes: https://tree.taiga.io/project/stoq/task/2077
Change-Id: Ibebe57906ada4127b58ca1a7b3f890f93affe0cf