Minor, refactor, remove unreacheable code
Reviewing the possible cases for non 200 return on TEF operations, there
was this case that would return success for a case that if reached was
actually an error (no ntk plugin) so keeping it would be wrong.
Since this code is already inside a block that checks for has_ntk, we
think it is safe to remove it and let a proper exception occur if for
some other reason ntk variable is not available.

Change-Id: Ie98b77dedefda0f0df9a21cf6bfab1542fb9652b
Fixes: https://gitlab.com/stoqtech/private/bdil/issues/97