Fix TEF and Error Messages
The event stream was locked because of a NTK library, so it was not
sending the TEF messages between the start and end of its process to the
user. The messages should be working fine now.

Error messages were also sent with incomplete data because of a flask
default setting that is now changed.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/132
Change-Id: Ib123d6e330024a630ffed91f62fa6c055788fe5a