Treat "sale already saved" errors
Use the same response of error printing coupon for when the error is one
of sale already save. This gives the users a chance to reprint the
coupon of the saved sale.

But only if a coupon was transmitted, if the coupon failed to be
transmitted the behavior will be the same as before this patch
( raise an error and a 500 response)

Change-Id: I3d98eeaf63257eb8c4bb017a4128120f55ac7c09
Fixes: https://gitlab.com/stoqtech/private/bdil/issues/142
Depends-On: I52e683ad372f233a0eb85dc33f9f7933ab104e25
Depends-On: I3c91c5ea4e5781b2a3d1a613b5642b585756171b