SAT device icon deactivated on NFe machines
The current problem is on NFe machines the SAT icon is following the
same behaviour as on SAT machines, which shows a red SAT icon when
the device is not connected.
This patch fixes it by always return 'OK' when there is no receivers for
SAT signal, what means no SAT plugin is enabled and should be a NFe
machine.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/200
Change-Id: I9e7a1cf94951d6ab5e633f3fcf21170aa7faf07b