Save authorizer name for tef payment
A TEF payment goes through an authorizer network and this information is
important for a later conciliation process.

We should save the name of the authorizer network so that we can use
that later.

Since we don't have an specific field for this info, lets use the card
payment device.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/79
Change-Id: Iefb9cf6aba68239483fe2909a96d7d399706ea8e