Support for generating debs of different branches
The release script gets the new version number based on a git show
command, but this command will return the last commit of the master
branch if jenkins is on a dettached head state.

This patch fix this by making a git checkout to the same branch as the
merged commit before checking the git show for version changes.

This patch also adds execution permission on all .sh scripts under the
bin folder.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/246
Change-Id: I204b44232b7d86152aa930c1b13a43f8719ce92b