REOPENED2 jenkins branch and master debs
The previous fix (with git checkout branch and git pull) is not good to
re-generate a gerrit triggered build with a jenkins retrigger. Because
it would pull the latest master changes, that can be ahead of the commit
we want to retrigger.

This patch changes from git pull to git fetch.

This patch also change the name for artifacts generated by manual builds
(alpha builds) to include a timestamp before the commit hash in their
version string.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/246
Change-Id: Ica139301d73b723b3a10ca3a5c39b983626c564f