Check if last till is open when confirming sale.
When a GET request is made on /till, it might return an error. The
frontend already handles this error, but, in case the frontend check
does not work, this commit handles the error in the backend. When the
user tries to confirm a sale, the backend checks if the last till is
open.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/282
Change-Id: I3c384e3de0f69ffb023dc0ec89242c5593f6ece6