Handle different card payments better.
The frontend might send the card type as an external plugin that adds a
new payment integration (that will be registered as a card payment).

Since the domain only allow for some basic types (credit/debit/voucher),
we need to fix the extraneous card type to credit.

Maybe we should handle this payment plugins differently in the future.o

Fixes: https://gitlab.com/stoqtech/stoq/issues/43
Change-Id: I363ab176f3134b167b96c252d54d33f5bffaf1b1