Make TEF summary default values work
Right now, the TEF summary has messages explaning
the default values for time and date inputs, but
this behavior doesn't work properly. In fact,
pressing only the enter button will make its flow
return a step.

This issue happens because of NTK flow, which needs
to be canceled if the green button is pressed but no
input was filled by the user. The opposite behavior
is expected for the SiTEF flow.

To fix that, the line of code who deals with
the NTK behavior is now on the project
stoq-plugin-ntk. For now, the server only
provides for the desired TEF plugin the user
input on StoqPDV, without changing it.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/526
Change-Id: Ib2071825ee83dfc4a66b7836f415a05b331a452e