Check if signal to generate till closing receipt had responses
If the plugin nonfiscal is not enabled on a stoq-server instance,
the signal GenerateTillClosingReceiptImageEvent will not get
answers, so we can't acess directly the value where the base64
image should be.

With this patch, we check if the signal had an answer. If it does,
we access and return the base64 image. If not, we return None.

Fixes: https://gitlab.com/stoqtech/private/stoq-server/issues/20
Change-Id: If42657ca3bf62ef4eb6caeb13b124a3dbc7a1d26