Force branch override using a parameter
Istead of using the hardcoded cnpj, use a paramter. To allow an easy
migration, the cnpj will work as a fallbakc for a while, until all
databases are configured properly.

Also, adding back code that turned out to be required after all.
Sellable does not provide an overide abstraction for default_sale_cfop.
This should be fixed with issue stoqtech/private/stoq-server#45

Fixes: https://gitlab.com/stoqtech/private/stoq-server/issues/44