Handle till already closed when closing till
Now that we have retries for closing the till, if the first request
fails, the second one will cause an error, since the till was already
closed by the first request.

The same thing when opening a till. If there is one already open, just
return true instead of raising an error

Fixes: https://gitlab.com/stoqtech/private/stoq-server/issues/59