Fix AssertionError on flask startup
The error was occurring because the workers always start to run before
any station is connected. As they try to send events to the stations
exceptions are raised and they stop working. This way, fixed the error
by ignoring any exception related to unconnected stations.

Moved the sending of the "device status changed" event to a method of
the EventStream module which allows to test this code more easily.

Fixes: https://gitlab.com/stoqtech/private/stoq-server/issues/60