Create a client address in sale post.
When a sale is being created, if there is a client, but the client is
missing an address, and there is an address in the payload, we should
create that.

This is usefull if the client already exisits in the database from a
previous sale (without address) and the caller is now sending an
address.