Fix B1Foods "payment method" aggregation
B1Foods consider card_data.card_type + card_data.provider as a payment method
we return this combination on the "nome" field, but the fields "codigo" and
"id" we were passing uuids for things we have a 1-1 match, like
payment_method or provider_id.

This causes a bug, when someone aggregates by "codigo", all MASTER payments
(credit + debit) gets summarized as the same "meiopagamento".

This patch changes what we use for code and id on the representation of a
"MeiosPagamento" item on both places where it is used, endpoints movimentocaixa
and meiospagamento.

See email thread.